+++ spec/blood_contracts/instrumentation/session_finalizer_spec.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+RSpec.describe BloodContracts::Instrumentation::SessionFinalizer do
+  before do
+  end
+
+  describe "defaults on new instance" do
+  end
+
+  describe "Basic finalizer" do
+  end
+
+  describe "Threads finalizer" do
+  end
+
+  describe "Fibers finalizer" do
+  end
+end